@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, Validators } from '@angular/forms';
 
 @Component({
     selector: 'app-forms-library',
@@ -24,18 +24,18 @@ export class FormsLibraryComponent implements OnInit {
 </mat-form-field>`;
 
     // Form
-    public form: FormGroup;
+    public form: UntypedFormGroup;
     public submitted = false;
     public hide: boolean = false;
 
     constructor(
-        private readonly _fb: FormBuilder
+        private readonly fb: UntypedFormBuilder
     ) {
         const formGroup = {
             email: ['', [Validators.required, Validators.email]],
             password: ['', Validators.required],
         };
-        this.form = this._fb.group(formGroup);
+        this.form = this.fb.group(formGroup);
     }
 
     ngOnInit() {
