@@ -1,33 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatDialogModule } from '@angular/material/dialog';
 import { MatExpansionModule } from '@angular/material/expansion';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatRadioModule } from '@angular/material/radio';
-import { MatSelectModule } from '@angular/material/select';
-import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatNativeDateModule } from '@angular/material/core';
-import { MatListModule } from '@angular/material/list';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatIconModule } from '@angular/material/icon';
 
 const ANGULAR_MATERIAL_MODULES = [
-    MatDialogModule,
     MatExpansionModule,
-    MatTooltipModule,
-    MatCheckboxModule,
-    MatRadioModule,
-    MatSelectModule,
-    MatDatepickerModule,
     MatFormFieldModule,
-    MatInputModule,
-    MatNativeDateModule,
-    MatListModule,
     MatTabsModule,
     MatMenuModule,
     MatSidenavModule,
@@ -40,6 +22,6 @@ const ANGULAR_MATERIAL_MODULES = [
         ...ANGULAR_MATERIAL_MODULES
     ],
     exports: [...ANGULAR_MATERIAL_MODULES],
-    providers: [MatDatepickerModule, MatSelectModule]
+    providers: []
 })
 export class AngularMaterialModule { }
