@@ -5,17 +5,17 @@ import { HeaderComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { ArrowScrollComponent } from './components/arrow-scroll/arrow-scroll.component';
 import { spinnerComponent } from './components/spinner/spinner.component';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { InfoboxComponent } from './components/infobox/infobox.component';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { LazyImgDirective } from './directives/lazy-img.directive';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 const components = [
-  HeaderComponent,
-  FooterComponent,
-  ArrowScrollComponent,
-  spinnerComponent,
-  InfoboxComponent
+    HeaderComponent,
+    FooterComponent,
+    ArrowScrollComponent,
+    spinnerComponent,
+    InfoboxComponent
 ];
 
 const material = [MatSnackBarModule, MatExpansionModule];
@@ -23,9 +23,9 @@ const material = [MatSnackBarModule, MatExpansionModule];
 const directive = [LazyImgDirective];
 
 @NgModule({
-  declarations: [...components, ...directive],
-  imports: [CommonModule, ReactiveFormsModule, FormsModule],
-  exports: [...components, ...material, ...directive],
-  providers: [],
+    declarations: [...components, ...directive],
+    imports: [CommonModule, ReactiveFormsModule, FormsModule],
+    exports: [...components, ...material, ...directive],
+    providers: [],
 })
-export class SharedModule {}
+export class SharedModule { }
